fix: parse Tabs to components package and remove from ui package
