+++ packages/ui/package.json
@@ -29,7 +29,6 @@
     "@theme-ui/match-media": "^0.17.1",
     "@ttoss/theme": "workspace:^",
     "react-select": "^5.9.0",
-    "react-tabs": "^6.1.0",
     "react-tooltip": "^5.28.0",
     "theme-ui": "^0.17.1"
   },
