+++ packages/ui/src/index.ts
@@ -41,7 +41,6 @@ export { Slider, type SliderProps } from './components/Slider';
 export { Spinner, type SpinnerProps } from './components/Spinner';
 export { Stack, type StackProps } from './components/Stack';
 export { Switch, type SwitchProps } from './components/Switch';
-export { Tabs } from './components/Tabs';
 export { Text, type TextProps } from './components/Text';
 export { Textarea, type TextareaProps } from './components/Textarea';
 export { Tooltip } from './components/Tooltip';
